+++ web/source/recorder_KeyboardScripts.js
@@ -1,7 +1,7 @@
 // JavaScript Document keyboardScripts.js: Keyboard management for KeymanWeb demonstration pages
 
 /* 
-    The keyboard name and/or ISO language code must be specified for each keyboard that is to be available.  
+    The keyboard name and/or BCP-47 language code must be specified for each keyboard that is to be available.
     If the same keyboard is used for several languages, it must be listed for each
     language, but the keyboard itself will only be loaded once. 
     If two (or more) keyboards are to be available for a given language, both must be listed.
@@ -11,12 +11,12 @@
   
     Each argument to addKeyboards() is a string, for example:
       european2         loads the current version of the Eurolatin 2 keyboard (for its default language)
-      european2@fra     loads the current version of the Eurolatin 2 keyboard for French
-      european2@fra@1.2 loads version 1.2 of the Eurolatin 2 keyboard for French
+      european2@fr      loads the current version of the Eurolatin 2 keyboard for French
+      european2@fr@1.2  loads version 1.2 of the Eurolatin 2 keyboard for French
       
     Argument syntax also supports the following extensions:
-      @fra              load the current version of the default keyboard for French
-      @fra$             load all available keyboards (current version) for French
+      @fr               load the current version of the default keyboard for French
+      @fr$              load all available keyboards (current version) for French
           
     Each call to addKeyboards() requires a single call to the remote server, 
     (unless all keyboards listed are local and fully specified) so it is better
@@ -45,26 +45,26 @@
     var kmw=keyman;
     
     // The first keyboard added will be the default keyboard for touch devices.
-    // For faster loading, it may be best for the default keybaord to be 
+    // For faster loading, it may be best for the default keyboard to be
     // locally sourced.
-    kmw.addKeyboards({id:'us',name:'English',languages:{id:'eng',name:'English'},
+    kmw.addKeyboards({id:'us',name:'English',languages:{id:'en',name:'English'},
       filename:'../testing/us-1.0.js'});
       
     // Add more keyboards to the language menu, by keyboard name,
-    // keyboard name and language code, or just the ISO 639 language code.  
+    // keyboard name and language code, or just the BCP-47 language code.
     // We use a different loading pattern here than in the samples version to provide a slightly different set of test cases.
-    kmw.addKeyboards('french','@heb');
-    kmw.addKeyboards({id:'european2', name:'EuroLatin2', languages: [{id:'nor'}, {id:'swe'}]}); // Loads from partial stub instead of the compact string.
+    kmw.addKeyboards('french','@he');
+    kmw.addKeyboards({id:'european2', name:'EuroLatin2', languages: [{id:'no'}, {id:'sv'}]}); // Loads from partial stub instead of the compact string.
   
     // Add a keyboard by language name.  Note that the name must be spelled
-    // correctly, or the keyboard will not be found.  (Using ISO codes is
+    // correctly, or the keyboard will not be found.  (Using BCP-47 codes is
     // usually easier.)
     kmw.addKeyboardsForLanguage('Dzongkha');
     
     // Add a fully-specified, locally-sourced, keyboard with custom font  
     kmw.addKeyboards({id:'lao_2008_basic',name:'Lao Basic',
       languages:{
-        id:'lao',name:'Lao',region:'Asia',
+        id:'lo',name:'Lao',region:'Asia',
         },
       filename:'../testing/lao_2008_basic.js'
       });   
@@ -73,7 +73,7 @@
     //  (a) a specific mapped input element input is focused, to ensure that the OSK appears
     //  (b) a specific keyboard is loaded, rather than the keyboard last used.         
   //window.setTimeout(function(){kmw.setActiveElement('ta1',true);},2500);
-  //window.setTimeout(function(){kmw.setActiveKeyboard('Keyboard_french','fra');},3000);
+  //window.setTimeout(function(){kmw.setActiveKeyboard('Keyboard_french','fr');},3000);
   
     // Note that locally specified keyboards will be listed before keyboards 
     // requested from the remote server by user interfaces that do not order
@@ -92,7 +92,11 @@
         break;
       case 2:
         sKbd=document.getElementById('kbd_id2').value.toLowerCase();
-        keyman.addKeyboards('@'+sKbd);
+        var rx=new RegExp(/^\w{2,3}$\$?/);
+        if(rx.test(sKbd))
+            keyman.addKeyboards('@'+sKbd);
+        else
+            alert('A BCP-47 language code must be exactly 2-3 letters long!');
         break;
       case 3:
         sKbd=document.getElementById('kbd_id3').value;